Convert download code into reusable function
