+++ web/pages/caregiver/resources/waivers.tsx
@@ -1,14 +1,18 @@
-import DownloadIcon from "@components/Icons/DownloadIcon";
-import BackButton from "@components/atoms/BackButton";
-import TitleTopBar from "@components/logos/TitleTopBar";
-import QuestionAnswer from "@components/resources/QuestionAnswer";
-import { BrowserWaiver } from "@lib/types/common";
-import { getCurrentCaregiver } from "db/actions/caregiver/Caregiver";
-import { getWaivers } from "db/actions/shared/Waiver";
-import { Timestamp } from "firebase-admin/firestore";
-import MarkdownIt from "markdown-it";
 import { GetServerSideProps, GetServerSidePropsContext } from "next";
 import { useRouter } from "next/router";
+import { Timestamp } from "firebase-admin/firestore";
+import MarkdownIt from "markdown-it";
+
+import { getCurrentCaregiver } from "db/actions/caregiver/Caregiver";
+import { getWaivers } from "db/actions/shared/Waiver";
+
+import { BrowserWaiver } from "@lib/types/common";
+import downloadWaiver from "@lib/utils/waiver";
+
+import BackButton from "@components/atoms/BackButton";
+import QuestionAnswer from "@components/resources/QuestionAnswer";
+import DownloadIcon from "@components/Icons/DownloadIcon";
+import TitleTopBar from "@components/logos/TitleTopBar";
 
 interface Props {
   waivers: BrowserWaiver[];
@@ -19,36 +23,6 @@ const mdParser = new MarkdownIt();
 export default function Waivers({ waivers }: Props) {
   const router = useRouter();
 
-  const handleDownload = async (waiver: BrowserWaiver) => {
-    const element = document.createElement("div");
-
-    // Add basic styling to the waiver PDF content
-    element.innerHTML = `
-    <div style="font-family: Arial, sans-serif; padding: 20px; max-width: 600px; margin: auto;">
-      <h1 style="text-align: center; font-size: 24px; color: #333; margin-bottom: 10px;">${waiver.name}</h1>
-      <div style="margin-bottom: 20px; font-size: 14px; color: #555;">
-        ${mdParser.render(waiver.content)}
-      </div>
-      <hr style="border: 1px solid #ddd; margin: 20px 0;">
-      <div style="display: flex; justify-content: space-between; font-size: 14px; color: #333;">
-        <span>Date: ${new Date(waiver.lastUpdated).toLocaleDateString()}</span>
-        <span>Signature: ${waiver?.caregiverName}</span>
-      </div>
-    </div>
-  `;
-
-    const html2pdf = (await import("html2pdf.js")).default;
-
-    html2pdf()
-      .set({
-        margin: [10, 10, 10, 10], // Adjust margins for better layout
-        filename: `${waiver.name}.pdf`,
-        html2canvas: { scale: 2 },
-        jsPDF: { orientation: "portrait" },
-      })
-      .from(element)
-      .save();
-  };
   return (
     <div className="w-full h-full flex flex-col justify-start items-start">
       <TitleTopBar title="Resources" />
@@ -81,7 +55,7 @@ export default function Waivers({ waivers }: Props) {
                     />
                     <button
                       className="flex gap-1 self-start items-center text-mbb-pink fill-mbb-pink font-semibold"
-                      onClick={() => handleDownload(waiver)}
+                      onClick={() => downloadWaiver(waiver)}
                     >
                       <span>Download Form</span> <DownloadIcon />
                     </button>
